+++ access_management/models/sale_order.py
@@ -40,9 +40,14 @@ class SaleOrder(models.Model):
             raise AccessError(_("Vous n'êtes pas autorisé à modifier des lignes de devis ou de commande."))
         return super().write(vals)
 
+    def _bd_manager_no_delete(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_bd_manager")
+
     def unlink(self):
         if self._dg_read_all_active():
             raise AccessError(_("Vous n'êtes pas autorisé à supprimer des lignes de devis ou de commande."))
+        if self._bd_manager_no_delete():
+            raise AccessError(_("Le groupe BD MANAGER n'est pas autorisé à supprimer des devis/commandes."))
         return super().unlink()
 
 
@@ -79,8 +84,13 @@ class SaleOrderLine(models.Model):
             raise AccessError(_("Vous n'êtes pas autorisé à modifier des lignes de devis ou de commande."))
         return super().write(vals)
 
+    def _bd_manager_no_delete(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_bd_manager")
+
     def unlink(self):
         if self._dg_read_all_active():
             raise AccessError(_("Vous n'êtes pas autorisé à supprimer des lignes de devis ou de commande."))
+        if self._bd_manager_no_delete():
+            raise AccessError(_("Le groupe BD MANAGER n'est pas autorisé à supprimer des lignes de devis/commande."))
         return super().unlink()
